+++ java/src/main/java/com/angrysurfer/midi/util/Constants.java
@@ -5,7 +5,7 @@ public class Constants {
     public static int DEFAULT_PPQ = 24;
     public static int DEFAULT_BAR_COUNT = 4;
     public static int DEFAULT_BEATS_PER_BAR = 4;
-    public static double DEFAULT_BEAT_DIVIDER = 4.0;
+    public static int DEFAULT_BEAT_DIVIDER = 4;
     public static float DEFAULT_BPM = 88;
     public static long DEFAULT_PART_LENGTH = 1L;
     public static int DEFAULT_MAX_TRACKS = 16;
